+++ spec/recipes/default_spec.rb
@@ -1,9 +1,11 @@
-# Encoding: UTF-8
+# encoding: utf-8
+# frozen_string_literal: true
 
 require_relative '../spec_helper'
 
 describe 'spotify::default' do
-  let(:runner) { ChefSpec::SoloRunner.new }
+  let(:platform) { { platform: 'ubuntu', version: '16.04' } }
+  let(:runner) { ChefSpec::SoloRunner.new(platform) }
   let(:chef_run) { runner.converge(described_recipe) }
 
   it 'installs the Spotify app' do