+++ spec/support/cookbooks/resource_spotify_app_test/recipes/remove.rb
@@ -1,5 +1,0 @@
-# Encoding: UTF-8
-
-spotify_app 'default' do
-  action :remove
-end