+++ spec/support/cookbooks/resource_test/metadata.rb
@@ -1,12 +1,12 @@
 # encoding: utf-8
 # frozen_string_literal: true
 
-name 'spotify_test'
+name 'resource_test'
 maintainer 'Jonathan Hartman'
 maintainer_email 'j@p4nt5.com'
 license 'apache2'
-description 'spotify_test'
-long_description 'spotify_test'
+description 'Resource test wrapper'
+long_description 'Resource test wrapper'
 version '0.0.1'
 
 depends 'spotify'