+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Spotify Cookbook CHANGELOG
 
 Unreleased
 ----------
+- Remove dependency on the apt cookbook
+- Remove dependency on the windows cookbook
+- Update dependency on the dmg cookbook
 
 v1.0.0 (2016-03-03)
 -------------------